% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial.R
\name{sig_buffer}
\alias{sig_buffer}
\title{Buffer GIS tool}
\usage{
sig_buffer(capa_entrada, distancia, dissolve = TRUE, campo = NA)
}
\arguments{
\item{capa_entrada}{Un objeto vectorial de la clase sf}

\item{distancia}{Distancia de vecindad para determinar la zona de influencia (m)}

\item{dissolve}{Parametro para controlar la fusion de geometrias}

\item{campo}{Parametro para fusionar por un atributo. Opcional}
}
\value{
Un objeto vectorial de la clase sfc_POLYGON
}
\description{
\code{sig_buffer} devuelve una zona de influencia entorno a las geometrias de entrada
}
\examples{
library(sf)
ciudades_data <- data.frame(
    nombre = c("Madrid", "Barcelona", "Sevilla"),
    poblacion = c(3.3, 1.6, 0.7),
    lon = c(-3.70, 2.17, -5.98),
    lat = c(40.41, 41.38, 37.38)
    )
ciudades_sf <- st_as_sf(ciudades_data, coords = c("lon", "lat"), crs = 4326)

sig_buffer(ciudades_sf, distancia=200)
sig_buffer(ciudades_sf, distancia=200, dissolve = TRUE)
sig_buffer(ciudades_sf, distancia=200, dissolve = FALSE)

}
